orders-service/requirements.txt
Django==4.2.3
psycopg2-binary==2.9.6
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.0
drf-yasg>=1.21.7
python-decouple>=3.8
gunicorn==20.1.0
django-filter
django-prometheus
django-cors-headers
whitenoise>=6.5.0
channels>=4.0
channels-redis>=4.0
redis==4.6.0
daphne
uvicorn
celery>=5.3
numpy>=1.26
pandas>=2.1